feat: add flask app & fix bugs
Add a basic flask app so that a UI can present the images returned from the Kempfolds app.

Fix the Kempfolds app so that it returns a random image. The issue was one of ordering in the calls within the app. It's a tactical fix and ultimately needs to be refactored to remove multiple calls to essentialliy get a complete object to work with

Cleanup

Remove setup, not ready yet